+++ Tone/core/AudioNode.js
@@ -1,5 +1,6 @@
 import Tone from "../core/Tone";
 import "../core/Context";
+import "../shim/AudioNode";
 
 /**
  *  @class Tone.AudioNode is the base class for classes which process audio.
@@ -184,15 +185,12 @@ Tone.AudioNode.prototype.connect = function(unit, outputNum, inputNum){
  */
 Tone.AudioNode.prototype.disconnect = function(destination, outputNum, inputNum){
 	if (Tone.isArray(this.output)){
-		if (Tone.isNumber(destination)){
-			this.output[destination].disconnect();
-		} else {
-			outputNum = Tone.defaultArg(outputNum, 0);
-			this.output[outputNum].disconnect(destination, 0, inputNum);
-		}
+		outputNum = Tone.defaultArg(outputNum, 0);
+		this.output[outputNum].disconnect(destination, 0, inputNum);
 	} else {
-		this.output.disconnect.apply(this.output, arguments);
+		Tone.disconnect(this.output, destination, outputNum, inputNum);
 	}
+	return this;
 };
 
 /**