+++ Tone/source/Oscillator.js
@@ -122,12 +122,18 @@ Tone.Oscillator.Type = {
 Tone.Oscillator.prototype._start = function(time){
 	this.log("start", time);
 	//new oscillator with previous values
-	this._oscillator = new Tone.OscillatorNode();
-	this._oscillator.setPeriodicWave(this._wave);
+	var oscillator = new Tone.OscillatorNode();
+	this._oscillator = oscillator;
+	if (this._wave){
+		this._oscillator.setPeriodicWave(this._wave);
+	} else {
+		this._oscillator.type = this._type;
+	}
 	//connect the control signal to the oscillator frequency & detune
 	this._oscillator.connect(this.output);
 	this.frequency.connect(this._oscillator.frequency);
 	this.detune.connect(this._oscillator.detune);
+
 	//start the oscillator
 	time = this.toSeconds(time);
 	this._oscillator.start(time);
@@ -214,11 +220,21 @@ Object.defineProperty(Tone.Oscillator.prototype, "type", {
 		return this._type;
 	},
 	set : function(type){
-		var coefs = this._getRealImaginary(type, this._phase);
-		var periodicWave = this.context.createPeriodicWave(coefs[0], coefs[1]);
-		this._wave = periodicWave;
-		if (this._oscillator !== null){
-			this._oscillator.setPeriodicWave(this._wave);
+		var isBasicType = [Tone.Oscillator.Type.Sine, Tone.Oscillator.Type.Square, Tone.Oscillator.Type.Triangle, Tone.Oscillator.Type.Sawtooth].includes(type);
+		if (this._phase === 0 && isBasicType){
+			this._wave = null;
+			this._partialCount = 0;
+			//just go with the basic approach
+			if (this._oscillator !== null){
+				this._oscillator.type = type;
+			}
+		} else {
+			var coefs = this._getRealImaginary(type, this._phase);
+			var periodicWave = this.context.createPeriodicWave(coefs[0], coefs[1]);
+			this._wave = periodicWave;
+			if (this._oscillator !== null){
+				this._oscillator.setPeriodicWave(this._wave);
+			}
 		}
 		this._type = type;
 	}
@@ -289,7 +305,7 @@ Object.defineProperty(Tone.Oscillator.prototype, "partialCount", {
  *  @returns {Object}
  */
 Tone.Oscillator.prototype.get = function(){
-	const values = Tone.prototype.get.apply(this, arguments);
+	var values = Tone.prototype.get.apply(this, arguments);
 	if (values.type !== Tone.Oscillator.Type.Custom){
 		delete values.partials;
 	}