+++ Tone/source/OscillatorNode.js
@@ -174,6 +174,13 @@ Tone.OscillatorNode.prototype.stop = function(time){
 		this._timeout = this.context.setTimeout(function(){
 			this._oscillator.stop(this.now());
 			this.onended();
+			//disconnect the object when it's ended
+			setTimeout(function(){
+				if (this._oscillator){
+					this._oscillator.disconnect();
+					this._gainNode.disconnect();
+				}
+			}.bind(this), 100);
 		}.bind(this), this._stopTime - this.context.currentTime);
 	} else {
 		//cancel the stop envelope
@@ -216,17 +223,20 @@ Object.defineProperty(Tone.OscillatorNode.prototype, "type", {
  *  @return  {Tone.OscillatorNode}  this
  */
 Tone.OscillatorNode.prototype.dispose = function(){
-	this.context.clearTimeout(this._timeout);
-	Tone.AudioNode.prototype.dispose.call(this);
-	this.onended = null;
-	this._oscillator.disconnect();
-	this._oscillator = null;
-	this._gainNode.dispose();
-	this._gainNode = null;
-	this.frequency.dispose();
-	this.frequency = null;
-	this.detune.dispose();
-	this.detune = null;
+	if (!this._wasDisposed){
+		this._wasDisposed = true;
+		this.context.clearTimeout(this._timeout);
+		Tone.AudioNode.prototype.dispose.call(this);
+		this.onended = null;
+		this._oscillator.disconnect();
+		this._oscillator = null;
+		this._gainNode.dispose();
+		this._gainNode = null;
+		this.frequency.dispose();
+		this.frequency = null;
+		this.detune.dispose();
+		this.detune = null;
+	}
 	return this;
 };